requirements.txt
sqlalchemy
alembic
pydantic
psycopg2-binary  # Driver PostgreSQL
python-dotenv    # Untuk keamanan credentials
flask
pyjwt
bcrypt
flasgger
pytest